Per client HTTP connection pooling
Implementation with static factory method
